Backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
boto3==1.36.8
botocore==1.36.8
click==8.1.8
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.7
h11==0.14.0
idna==3.10
jmespath==1.0.1
mangum==0.19.0
pip-autoremove==0.10.0
pip-chill==1.0.3
pydantic==2.10.6
pydantic-settings==2.1.0
pydantic_core==2.27.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.11.2
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
typing_extensions==4.12.2
urllib3==1.26.20
uvicorn==0.34.0
groq==0.18.0
